+++ rpc-filter/Cargo.toml
@@ -1,0 +1,17 @@
+[package]
+name = "solders-rpc-filter"
+version = "0.14.4"
+edition = "2021"
+include = ["/src"]
+license = "MIT"
+repository = "https://github.com/kevinheavey/solders"
+
+[dependencies]
+pyo3 = { workspace = true }
+serde = { workspace = true }
+derive_more = { workspace = true }
+solders-macros = { workspace = true }
+solders-traits = { workspace = true }
+thiserror = { workspace = true }
+bs58 = { workspace = true }
+base64 = { workspace = true }
